refactor(tools): split ReadPdf and ReadHtml out of Read tool
The Read tool had PDF extraction and HTML conversion baked in,
violating SRP and polluting its schema with a `pages` parameter
that caused LLM agents to pass empty strings on non-PDF files,
triggering repeated failures.

- New `loopal-tool-read-pdf` crate: ReadPdf tool with `pages` param
- New `loopal-tool-read-html` crate: ReadHtml tool for HTML→text
- Read tool now only handles text files (clean schema, no `pages`)
