+++ crates/tools/filesystem/read-html/src/lib.rs
@@ -1,0 +1,3 @@
+mod tool;
+
+pub use tool::ReadHtmlTool;
