+++ crates/tools/filesystem/read-html/tests/suite.rs
@@ -1,0 +1,2 @@
+#[path = "suite/read_html_tool_test.rs"]
+mod read_html_tool_test;
