+++ crates/tools/filesystem/read-pdf/BUILD.bazel
@@ -1,29 +1,30 @@
 load("@rules_rust//rust:defs.bzl", "rust_library", "rust_test")
 
 rust_library(
-    name = "loopal-tool-read",
+    name = "loopal-tool-read-pdf",
     srcs = glob(["src/**/*.rs"]),
     edition = "2024",
     visibility = ["//visibility:public"],
     deps = [
         "//crates/loopal-error:loopal-error",
         "//crates/loopal-tool-api:loopal-tool-api",
+        "@crates//:pdf-extract",
         "@crates//:serde_json",
     ],
     proc_macro_deps = ["@crates//:async-trait"],
 )
 
 rust_test(
-    name = "loopal-tool-read-test",
+    name = "loopal-tool-read-pdf-test",
     srcs = glob(["tests/**/*.rs"]),
     crate_root = "tests/suite.rs",
     edition = "2024",
     deps = [
-        ":loopal-tool-read",
+        ":loopal-tool-read-pdf",
         "//crates/loopal-backend:loopal-backend",
+        "//crates/loopal-tool-api",
+        "@crates//:serde_json",
         "@crates//:tempfile",
         "@crates//:tokio",
-        "@crates//:serde_json",
-        "//crates/loopal-tool-api",
     ],
 )
