+++ crates/tools/filesystem/read-pdf/src/lib.rs
@@ -1,0 +1,5 @@
+mod page_range;
+mod tool;
+
+pub use page_range::parse_page_range;
+pub use tool::ReadPdfTool;
