+++ crates/tools/filesystem/read-pdf/tests/suite.rs
@@ -1,0 +1,4 @@
+#[path = "suite/page_range_test.rs"]
+mod page_range_test;
+#[path = "suite/read_pdf_tool_test.rs"]
+mod read_pdf_tool_test;
