+++ crates/tools/filesystem/read/tests/suite.rs
@@ -1,9 +1,4 @@
-// Single test binary — includes all test modules
-#[path = "suite/read_html_test.rs"]
-mod read_html_test;
-#[path = "suite/read_pdf_test.rs"]
-mod read_pdf_test;
-#[path = "suite/read_tool_edge_test.rs"]
-mod read_tool_edge_test;
 #[path = "suite/read_tool_test.rs"]
 mod read_tool_test;
+#[path = "suite/read_tool_edge_test.rs"]
+mod read_tool_edge_test;
